requirements.txt
Django
django-bootstrap4
uwsgi
psycopg2
numpy
Pillow
matplotlib
opencv-python
gluoncv
scipy
scikit-image
django-cleanup
django-stdimage
tensorflow
tensorflowjs
tqdm